Simpler stack to fix #42
