+++ .github/workflows/main.yaml
@@ -20,4 +20,4 @@ jobs:
 
       - run: bun install
       - run: bun lint
-      - run: bun vitest
+      - run: bun test
