+++ src/QDL/saharaDefs.js
@@ -1,4 +1,4 @@
-import { structHelper_io } from "./utils"
+import { structHelper_io } from "./utils.js"
 
 
 export const cmd_t = {
