+++ src/utils/image.js
@@ -6,7 +6,7 @@ export function useImageWorker() {
   const apiRef = useRef()
 
   useEffect(() => {
-    const worker = new Worker(new URL('../workers/image.worker', import.meta.url), {
+    const worker = new Worker('src/workers/image.worker.js', {
       type: 'module',
     })
     apiRef.current = Comlink.wrap(worker)
